# Conflicts:
#	composer.json
#	composer.lock

+++ app/code/Magento/Catalog/Controller/Adminhtml/Product/Set/Delete.php
@@ -6,6 +6,8 @@
  */
 namespace Magento\Catalog\Controller\Adminhtml\Product\Set;
 
+use Magento\Framework\Exception\NotFoundException;
+
 class Delete extends \Magento\Catalog\Controller\Adminhtml\Product\Set
 {
     /**
@@ -29,19 +31,25 @@ class Delete extends \Magento\Catalog\Controller\Adminhtml\Product\Set
 
     /**
      * @return \Magento\Backend\Model\View\Result\Redirect
+     * @throws NotFoundException
      */
     public function execute()
     {
-        $setId = $this->getRequest()->getParam('id');
+        if (!$this->getRequest()->isPost()) {
+            throw new NotFoundException(__('Page not found'));
+        }
+
+        $setId = (int)$this->getRequest()->getParam('id');
         $resultRedirect = $this->resultRedirectFactory->create();
         try {
             $this->attributeSetRepository->deleteById($setId);
-            $this->messageManager->addSuccess(__('The attribute set has been removed.'));
+            $this->messageManager->addSuccessMessage(__('The attribute set has been removed.'));
             $resultRedirect->setPath('catalog/*/');
         } catch (\Exception $e) {
-            $this->messageManager->addError(__('We can\'t delete this set right now.'));
+            $this->messageManager->addErrorMessage(__('We can\'t delete this set right now.'));
             $resultRedirect->setUrl($this->_redirect->getRedirectUrl($this->getUrl('*')));
         }
+
         return $resultRedirect;
     }
 }
